fix(Friend) : 친구 신청 검증 로직 변경
